third_party/WebKit/LayoutTests/http/tests/devtools/network/network-xhr-data-received-async-response-type-blob-expected.txt

CONSOLE MESSAGE: line 5: XHR loaded: 1
Tests that dataReceived is called on NetworkDispatcher for XHR with responseType="blob".

Received data for resource.php
SUCCESS
